fix: cross-reference pass 2 — API correctness for 40+ modules (MC 1.21.11)
Reach:
- Complete rewrite: now uses EntityAttributes.ENTITY_INTERACTION_RANGE and
  BLOCK_INTERACTION_RANGE (setBaseValue) instead of a manual entity loop that
  missed block reach entirely. Verified attribute names via javap on 1.21.11 jar.
  Default baselines: entity=3.0, block=4.5.

Movement modules:
- Spider: addSetting → addNumber (slider), remove stale parseInt helper
- Jesus: addSetting("Lava","false") → addBool("Lava", false) for proper toggle UI
- Phase: addSetting → addNumber (slider), remove stale parseInt helper
- ElytraFlight: remove stale parseDouble helper; confirmed ClientCommandC2SPacket
  and isGliding() API are correct in 1.21.11

Verified correct (no changes needed):
- NoClip: player.noClip is public in 1.21.11 (confirmed via javap)
- BowAimbot: Entity.lastX/Y/Z are public, CrossbowItem.isCharged() unchanged
- AutoArmor: already uses 1.21.x DataComponentTypes/EquippableComponent correctly
- AutoFarm/AutoFish/Nuker: attackBlock/interactBlock signatures unchanged
- AutoSwitch: getMiningSpeedMultiplier(state) correct in 1.21.11
- KillauraLegit: smooth rotation and canSee() LOS check correct
- AimAssist: MathHelper.wrapDegrees API unchanged
- Timer: getSpeed() pattern correct for mixin use

Systemic fix — all modules:
- parseInt helpers: replace Integer.parseInt(s) with (int) Double.parseDouble(s)
  to handle "4.0" strings that addNumber emits (was silently failing at runtime)
- StorageESP, ESP, Nametags: addSetting("x","true/false") → addBool for
  proper toggle rendering in ClickGUI

All 160 modules compile cleanly on MC 1.21.11.

+++ src/main/java/com/claudemc/module/impl/combat/TriggerBot.java
@@ -56,6 +56,6 @@ public class TriggerBot extends Module {
     }
 
     private int parseInt(String s, int def) {
-        try { return Integer.parseInt(s.trim()); } catch (Exception ex) { return def; }
+        try { return (int) Double.parseDouble(s.trim()); } catch (Exception ex) { return def; }
     }
 }
