+++ src/main/java/com/claudemc/module/impl/misc/AutoAuth.java
@@ -71,6 +71,6 @@ public class AutoAuth extends Module {
     }
 
     private int parseInt(String s, int def) {
-        try { return Integer.parseInt(s.trim()); } catch (Exception e) { return def; }
+        try { return (int) Double.parseDouble(s.trim()); } catch (Exception e) { return def; }
     }
 }
