+++ src/main/java/com/claudemc/module/impl/misc/MassTPA.java
@@ -73,6 +73,6 @@ public class MassTPA extends Module {
     }
 
     private int parseInt(String s, int d) {
-        try { return Integer.parseInt(s); } catch (Exception e) { return d; }
+        try { return (int) Double.parseDouble(s); } catch (Exception e) { return d; }
     }
 }
