+++ src/main/java/com/claudemc/module/impl/misc/ServerFinder.java
@@ -560,7 +560,7 @@ public class ServerFinder extends Module {
     }
 
     private int parseInt(String s, int def) {
-        try { return Integer.parseInt(s.trim()); } catch (Exception e) { return def; }
+        try { return (int) Double.parseDouble(s.trim()); } catch (Exception e) { return def; }
     }
 
     private static class ServerEntry {
