+++ src/main/java/com/claudemc/module/impl/render/BlockScanModule.java
@@ -78,6 +78,6 @@ public abstract class BlockScanModule extends Module {
     protected abstract void scan(MinecraftClient client, List<Highlight> out);
 
     protected static int parseInt(String s, int d) {
-        try { return Integer.parseInt(s); } catch (Exception e) { return d; }
+        try { return (int) Double.parseDouble(s); } catch (Exception e) { return d; }
     }
 }
