+++ src/main/java/com/claudemc/module/impl/render/Breadcrumbs.java
@@ -66,7 +66,7 @@ public class Breadcrumbs extends Module {
     }
 
     private int parseInt(String s, int def) {
-        try { return Integer.parseInt(s.trim()); } catch (Exception e) { return def; }
+        try { return (int) Double.parseDouble(s.trim()); } catch (Exception e) { return def; }
     }
 
     private double parseDouble(String s, double def) {
