+++ src/main/java/com/claudemc/module/impl/render/PlayerFinder.java
@@ -65,6 +65,6 @@ public class PlayerFinder extends Module {
     }
 
     private static int parseInt(String s, int def) {
-        try { return Integer.parseInt(s.trim()); } catch (Exception e) { return def; }
+        try { return (int) Double.parseDouble(s.trim()); } catch (Exception e) { return def; }
     }
 }
