+++ src/main/java/com/claudemc/module/impl/render/TimeChanger.java
@@ -40,6 +40,6 @@ public class TimeChanger extends Module {
     }
 
     private int parseInt(String s, int def) {
-        try { return Integer.parseInt(s.trim()); } catch (Exception e) { return def; }
+        try { return (int) Double.parseDouble(s.trim()); } catch (Exception e) { return def; }
     }
 }
