+++ src/main/java/com/claudemc/module/impl/world/Tillaura.java
@@ -65,6 +65,6 @@ public class Tillaura extends Module {
     }
 
     private int parseInt(String s, int d) {
-        try { return Integer.parseInt(s); } catch (Exception e) { return d; }
+        try { return (int) Double.parseDouble(s); } catch (Exception e) { return d; }
     }
 }
