+++ src/main/java/com/claudemc/module/impl/world/VeinMiner.java
@@ -48,6 +48,6 @@ public class VeinMiner extends Module {
     }
 
     private int parseInt(String s, int d) {
-        try { return Integer.parseInt(s); } catch (Exception e) { return d; }
+        try { return (int) Double.parseDouble(s); } catch (Exception e) { return d; }
     }
 }
